+++ examples/src/main/scala/com/github/mjakubowski84/parquet4s/akka/WriteAndReadCustomTypeAkkaApp.scala
@@ -12,7 +12,7 @@ object WriteAndReadCustomTypeAkkaApp extends App {
   object Data {
     def generate(count: Int): Iterator[Data] = Iterator.range(1, count).map(i => Data(id = i, dict = Dict.random))
   }
-  case class Data(id: Long, dict: Dict.Type)
+  case class Data(id: Int, dict: Dict.Type)
 
   val data = () => Data.generate(count = 100)
   val path = Path(Files.createTempDirectory("example"))
